Facial_Composite/requirements.txt
huggingface-hub==0.31.1
transformers==4.49.0
diffusers==0.32.2
bitsandbytes==0.45.5
accelerate==1.5.2
protobuf==6.30.1
optimum-quanto==0.2.7
sentencepiece==0.2.0
gradio==5.29.1
opencv-python==4.11.0.86